requirements.txt
# Dependencies maintained by OCCRP
banal==1.0.1
urlnormalizer==1.2.5
https://github.com/malteos/followthemoney/archive/qurator.zip#egg=followthemoney
followthemoney-store[postgresql]==2.2.1
fingerprints==1.0.1
servicelayer[google,amazon]==1.13.5
normality==2.1.1
pantomime==0.4.0

# Flask ecosystem
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Cors==3.0.8
Flask-Babel==1.0.0
flask-talisman==0.7.0
SQLAlchemy==1.3.18
alembic==1.4.2
authlib==0.14.3

elasticsearch==7.8.1
marshmallow==2.19.2
gunicorn[eventlet]==20.0.4
jsonschema==3.2.0
apispec==3.3.1
apispec-webframeworks==0.5.2
blinker==1.4
Babel==2.8.0
PyYAML==5.3.1
python-frontmatter==0.5.0
pyjwt==1.7.1
cryptography==3.0
rdflib==5.0.0
requests[security]==2.24.0
urllib3==1.25.10
networkx==2.4
zipstream==1.1.4
tabulate==0.8.7

# Testing dependencies
factory-boy==2.12.0
Faker==4.1.1
nose
coverage
coveralls

# Tracing and logging libraries
python-json-logger==0.1.11
google-cloud-logging==1.15.0

# Custom dependencies
spacy==2.3.2
